requirements.txt
# Proposal Agent — RFP → Impact-8 PPTX 제안서 자동 생성
# Python 3.10+

# Document Parsing (PDF, DOCX, TXT, PPTX)
pypdf>=3.17.0
pdfplumber>=0.10.0
python-docx>=1.1.0

# PPTX Generation
python-pptx>=0.6.23

# AI/LLM (.env LLM_PROVIDER: claude | gemini | groq 중 1개 선택)
anthropic>=0.39.0
google-genai>=1.0.0
groq>=0.4.0

# Data Validation
pydantic>=2.5.0

# CLI & UI
typer>=0.9.0
rich>=13.7.0

# Utilities
python-dotenv>=1.0.0
loguru>=0.7.0

# Testing
pytest>=7.0.0

# Optional: 차트를 이미지로 삽입할 때 사용 (현재는 python-pptx 셰이프 기반 렌더링)
# matplotlib>=3.8.0
# Pillow>=10.0.0
